Different pages depending on user type User can declare himself as positive Admin can set user pcr result as negative Admin can delete other users

+++ src/main/java/com/covid19/Servlet/AdminPanelServlet.java
@@ -23,13 +23,26 @@ public class AdminPanelServlet extends HttpServlet {
 
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         String deleteUser = request.getParameter("deleteuser");
+        String negUser = request.getParameter("neguser");
 
-        if (deleteUser != ""){
+        if (deleteUser != "" && negUser == ""){
             try {
                 if(dbSetup.deleteUser(deleteUser)){
                     request.setAttribute("messagedelete","*User Deleted");
                 } else {
                     request.setAttribute("messagedelete","*User Not found");
+                }
+            } catch (ClassNotFoundException e) {
+                e.printStackTrace();
+            }
+        }
+
+        if (negUser != "" && deleteUser == ""){
+            try {
+                if(dbSetup.setNegative(negUser)>0){
+                request.setAttribute("messagenegative","*User set as Negative");
+                } else {
+                    request.setAttribute("messagenegative","*An error has occured");
                 }
             } catch (ClassNotFoundException e) {
                 e.printStackTrace();
